backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.3
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2